+++ app/src/actions/post.lua
@@ -54,6 +54,8 @@ return {
 		self.comments, self.pagination =
 			paginate_thread(Comments:thread(post_data.id), self.params.page, COMMENTS_PER_PAGE)
 
+		self.tags = require("src.models.tags"):for_post(post_data.id)
+
 		-- print("Post data:")
 		-- require 'pl.pretty'.dump(post_data[1])